requirements/app.txt
galaxy.plugin.api==0.65
psutil==5.6.6
aiofiles==0.4.0